Updated vecManager - new interface for parsing format string for each of the vector generation methods. The nomenclature has changed to
writeXXVec() -> genXXVec()

where writeXXVec() is a frontend function for each of the vector generators which
does some checking and clean up, then calls genXXVec() to create the actual test
vector. The data is written by a seperate function vecDiskWrite() which is called from
within writeXXVec(). Parsing is seperated out into parseFmt() method in vecManager.

Still todo:

- Write set of test for vector manager
- Finish sparse buffer module, and check this works properly with vecManager class
- Perform overall test

Proposed new features:

- Some kind of object that can do ML for segmentation. Not completely sure yet whether
this should be a subclass of cSegmenter or a seperate thing. In any case, it makes the
most sense for csSegmenter to call methods in this class to perform the segmentation,
but it seems like there should be some kind of trick that lets you only create this
ML object when you need it, get it to give you back some data, and then destroy the
object, saving whatever the memory overhead for ML until its actually required.

This module will contain methods for any of the ML methods I want to try for
segmenting stuff (and particuarly small stuff). Some ideas:

- PCA - Should map well to hardware, as long as training offline
- HMM - Seems like hardware can do this as long as training is offline, but I don't
        know if this will give an output that camshift can accept
- NN  - Maybe a (smallish) Neural Net will be suitable..?

The point of an ML class is obviously to test that these methods can provide good
segmentation performance with relevant data output for this application, as well as
assess the memory requirments, etc.

Note that the above is not in this checkin