IMRECT DRAMA New method to specify imrect by pressing 'r' key twice.
PROBLEM: Currently, the rectangle steadfastly refuses to show in the specified
axes. This has been a problem since the imrect method was brought into the project.
Note that in all the test and examples I've run to date, the imrect functionality
works as expected. It's only in csToolGUI that issues appear (perhaps there needs to
be some kind of programmatic ordering of the GUI elements so that the imrect handles
appear on top of the image)?

Overall, number of changes in this commit are minor. Mostly small edits to GUI. The
largest architectural change is with the sub-GUI menu items which now pass arguments
back to the main GUI in the correct way (although its still broken). These are also
not reading the argument data correctly.

Need to get that vecManager in there soon