Nearly got tracker going. Need to check that window parameters are passed correctly from one frame to the next. The operations seem to occur in a strange order.
For each frame, the window should stay fixed until the final iteration. Within the
window, the accumulator should find the 'average' position of the pixels (centroid).
At the end of the frame, once the mean has converged, translate the tracking window
over the new mean, and repeat.

At the moment, the centroid seems to not move very much inside the window. Also, it
seems (by inspection) that the window moves in the opposite direction that it ought
to. (This is probably some irritating issue with how MATLAb defines the origin of the
image, so look into this)

Started a new parameter browser to interactively browse the frame paramters for each
iteration in each frame. The panel for this should have a small window where the
frame can be previewed (in a more advanced GUI, this might be thumbnaield, but in this
it'll probably just be the same image on a small axis)