Just about to modify the trackFrame() routine.
This will be split into two methods - trackFrame(), which is a wrapper call in
csTracker class, and another meanshift loop method. The point of this change is to
claify some architectural issues in csTool (ie: where and how are tracking windows
created/destroyed), and to split the source files to make editing easier.